Login working plus json

+++ src/main/java/org/spe/biologicaldata/webapplication/service/Database.java
@@ -2,15 +2,15 @@ package org.spe.biologicaldata.webapplication.service;
 
 import org.spe.biologicaldata.webapplication.model.Image;
 import org.spe.biologicaldata.webapplication.repository.ImageRepository;
+import org.spe.biologicaldata.webapplication.wrapper.ImagePathWrapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.io.File;
 import java.util.*;
 
 @Service
-public class Database implements DatabaseController {
+public class Database implements DatabaseService {
 
     private StorageService storageService;
     private ImageRepository imageRepository;
@@ -19,47 +19,43 @@ public class Database implements DatabaseController {
     public Database(StorageService storageService, ImageRepository imageRepository) {
         this.storageService = storageService;
         this.imageRepository = imageRepository;
+    }
 
-        //Load files from static/images into the in memory database
-        File folder = new File("src/main/resources/static/images");
-        List<File> files;
-        try{
-            files = Arrays.asList(Objects.requireNonNull(folder.listFiles()));
-        } catch(NullPointerException e) {
-            boolean bool = folder.mkdir();
-            files = Arrays.asList(Objects.requireNonNull(folder.listFiles()));
+    @Override
+    public Boolean storeImage(Image imageInfo, MultipartFile image, Boolean enablePublicReadAccess) {
+        Optional<ImagePathWrapper> pathUrl = storageService.store(image, enablePublicReadAccess);
+        if(pathUrl.isPresent()) {
+            imageInfo.setImageUrl(pathUrl.get().getLink());
+            imageInfo.setBucketPathId(pathUrl.get().getPathId());
+            imageRepository.save(imageInfo);
+            return true;
         }
+        return false;
+    }
 
-        for (File file : files) {
-            if (file.isFile()) {
-                String path = file.getPath();
-                imageRepository.save(new Image(path.substring(path.indexOf("images")),file.getName(),"","","",""));
-            } else if (file.isDirectory()) {
-                files.addAll(Arrays.asList(Objects.requireNonNull(file.listFiles())));
-            }
+    @Override
+    public Boolean storeImage(MultipartFile image, Boolean enablePublicReadAccess) {
+        Optional<ImagePathWrapper> pathUrl = storageService.store(image, enablePublicReadAccess);
+        if(pathUrl.isPresent()) {
+            Image imageRow = new Image(pathUrl.get().getLink(), pathUrl.get().getPathId(),
+                    image.getName(), "", "", "", "");
+            imageRepository.save(imageRow);
+            return  true;
         }
+        return false;
     }
 
     @Override
-    public void storeImage(Image imageInfo, MultipartFile image) {
-        String pathUrl = storageService.store(image, true);
-        imageInfo.setImageUrl(pathUrl);
-        imageRepository.save(imageInfo);
-    }
-
-    @Override
-    public void storeImage(MultipartFile image) {
-        String pathUrl = storageService.store(image, true);
-        Image imageRow = new Image(pathUrl, image.getName(), "", "", "", "");
-        imageRepository.save(imageRow);
-    }
-
-    @Override
-    public void storeImage(String title, String author, String writtenDate,
-                           String page, String description, MultipartFile image) {
-        String pathUrl = storageService.store(image, true);
-        Image imageRow = new Image(pathUrl, title, author, writtenDate, page, description);
-        imageRepository.save(imageRow);
+    public Boolean storeImage(String title, String author, String writtenDate,
+                              String page, String description, MultipartFile image, Boolean enablePublicReadAccess) {
+        Optional<ImagePathWrapper> pathUrl = storageService.store(image, enablePublicReadAccess);
+        if(pathUrl.isPresent()){
+            Image imageRow = new Image(pathUrl.get().getLink(), pathUrl.get().getPathId(),
+                    title, author, writtenDate, page, description);
+            imageRepository.save(imageRow);
+            return true;
+        }
+     return false;
     }
 
     @Override
@@ -73,25 +69,25 @@ public class Database implements DatabaseController {
     }
 
     @Override
-    public void deleteImageById(Long id) {
+    public Boolean deleteImageById(Long id) {
         Optional<Image> image= imageRepository.findById(id);
         if(image.isPresent()) {
-            storageService.delete(image.get().getImageUrl());
+            storageService.delete(image.get().getBucketPathId());
             imageRepository.deleteById(id);
+            return true;
         }
+        return false;
     }
 
     @Override
-    public void deleteImage(Image image) {
-        storageService.delete(image.getImageUrl());
-        imageRepository.delete(image);
+    public Boolean deleteImage(Image image) {
+        if(!storageService.delete(image.getBucketPathId()))
+            return false;
+        if(imageRepository.existsById(image.getId())) {
+            imageRepository.delete(image);
+            return true;
+        }
+        return false;
     }
 
-    @Override
-    public void deleteAll() {
-        List<Image> images = new ArrayList<>((Collection<Image>) imageRepository.findAll());
-        for(Image image : images) {
-            deleteImage(image);
-        }
-    }
 }
